css file referenced
This file should be in the same folder as both the africanstorybooks.htmlf and the index.html (so there should be two copies). Otherwise edit the reference to this file from "index..." to "/index.." to use the file in the root directory. Having a copy in the folder allows it to be standalone